Added bzip2 support, enabled using scramble -j.
It was included in the CRAM 1.0 spec but removed from 2.0. I added it
to investigate how well it performs. Conclusion: 5-10% space
improvement,
but something like twice as slow. I don't think it's worth the
overhead and long-term we should go to modelling + arithmetic coding
to get better time/size tradeoffs.


git-svn-id: svn://svn.code.sf.net/p/staden/code/io_lib@3292 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1