Bug fix for reference handling when fetching via MD5/http.
Firstly the fd->refs->lock mutex has been moved to just before
cram_populate_ref instead of just after, given that this code is not
thread safe (it potentially uses open_path_url+curl in a manner which
shares the curl handle between threads).

Secondly certain thread orderings could cause the reference count to
hit zero and then try to reaccess it again, which triggered a
bug. ref->length is now reset to zero as well as ->seq=NULL as both
are checked.


git-svn-id: svn://svn.code.sf.net/p/staden/code/io_lib@3467 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1