Add mfsteal to mFILE, and use it in cram_populate_ref.
mfsteal returns the data stored in an mFILE.  The mFILE itself is closed
after the data has been detached.  This can be used to replace an
unnecessary allocation and copy if the entire contents of the mFILE are
wanted.

Update cram_populate_ref to use mfsteal.  This halves the memory it uses
when trying to load via REF_PATH.



git-svn-id: svn://svn.code.sf.net/p/staden/code/io_lib@3533 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1