Added BAM_CONSUME_REF(op) and BAM_CONSUME_SEQ(op) macros to help catch all types of cigar operations that step along the reference and/or sequence.
Added uses of these in bam_construct_seq() which wasn't correctly
handling the X and = cigar codes (not that we ever generate these in
our own code, but someone else may be using this as a library).

Finally fixed a bug in the interpretation of beg,end for the bin
calculation.  It was being used as [beg,end] instead of [beg,end),
causing bins to change in specific locations.


git-svn-id: svn://svn.code.sf.net/p/staden/code/io_lib@3731 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1